refactor: 기존 구현 v1으로 이동
