+++ src/test/java/nextstep/ladder/model/v1/OneByTwoDrawStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.model;
+package nextstep.ladder.model.v1;
 
 import nextstep.ladder.model.v1.DrawStrategy;
